requirements.txt
langchain
openai
pandas
openpyxl
streamlit
python-dotenv
